+++ packages/@kahi-ui-docs-kit/src/shared/components/shell/ShellOmni.svelte
@@ -42,7 +42,6 @@
                   on_click: on_theme_click.bind(null, _htmlpalette, _prefersdark),
               },
           ];
-
 </script>
 
 <Omni.Container