+++ packages/@kahi-ui-docs/src/components/repl/REPLLoadingHero.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
     import {Stack, Wave} from "@kahi-ui/framework";
-
 </script>
 
 <Stack alignment="center" height="viewport-100">