+++ packages/@kahi-ui-docs/src/routes/index.svelte
@@ -10,7 +10,6 @@
     import LandingHero from "../components/landing/LandingHero.svelte";
 
     import SnippetREPL from "../components/snippets/SnippetREPL.svelte";
-
 </script>
 
 <LandingHero />
@@ -23,7 +22,11 @@
         <Text is="small">Develop more UI without writing essays worth of markup.</Text>
     </Text>
 
-    <SnippetREPL documentation="samples/landing" snippet="hero-preview" rotation="horizontal" />
+    <SnippetREPL
+        documentation="samples/landing-page"
+        snippet="landing-preview"
+        rotation="horizontal"
+    />
 
     <Divider palette="accent" margin_y="huge" />
 
@@ -34,7 +37,10 @@
         <Text is="small">Opinonated when needed. Flexible when wanted.</Text>
     </Text>
 
-    <Grid points={["2", "mobile:1"]} spacing={["large", "tablet:medium", "mobile:medium"]}>
+    <Grid.Container
+        points={["2", "mobile:1"]}
+        spacing={["large", "tablet:medium", "mobile:medium"]}
+    >
         <LandingCard icon={GridIcon} heading="Composable">
             Designed with <Text is="strong">Composability</Text> in mind, author new Components easily.
         </LandingCard>
@@ -57,7 +63,7 @@
             </Text>
             enhancing the <Text is="strong">UX</Text>.
         </LandingCard>
-    </Grid>
+    </Grid.Container>
 </Container>
 
 <LandingFooter />