+++ src/components/display/list/index.ts
@@ -1,3 +1,2 @@
+export {default as Container} from "./ListContainer.svelte";
 export {default as Item} from "./ListItem.svelte";
-export {default as Ordered} from "./OrderedList.svelte";
-export {default as Unordered} from "./UnorderedList.svelte";