+++ src/components/layouts/grid/stories/GridPointsStory.svelte
@@ -3,8 +3,7 @@
     import Stack from "../../stack/Stack.svelte";
     import Text from "../../../typography/text/Text.svelte";
 
-    import Grid from "../Grid.svelte";
-
+    import * as Grid from "../index";
 </script>
 
 <Stack orientation="horizontal" spacing="medium" variation="wrap">
@@ -12,7 +11,7 @@
         <Text is="strong">12</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="12" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="12" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -28,7 +27,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -36,7 +35,7 @@
         <Text is="strong">11</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="11" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="11" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -52,7 +51,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -60,7 +59,7 @@
         <Text is="strong">10</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="10" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="10" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -76,7 +75,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -84,7 +83,7 @@
         <Text is="strong">9</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="9" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="9" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -100,7 +99,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -108,7 +107,7 @@
         <Text is="strong">8</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="8" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="8" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -124,7 +123,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -132,7 +131,7 @@
         <Text is="strong">7</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="7" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="7" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -148,7 +147,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -156,7 +155,7 @@
         <Text is="strong">6</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="6" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="6" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -172,7 +171,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -180,7 +179,7 @@
         <Text is="strong">5</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="5" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="5" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -196,7 +195,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -204,7 +203,7 @@
         <Text is="strong">4</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="4" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="4" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -220,7 +219,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -228,7 +227,7 @@
         <Text is="strong">3</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="3" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="3" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -244,7 +243,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -252,7 +251,7 @@
         <Text is="strong">2</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="2" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="2" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -268,7 +267,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 
@@ -276,7 +275,7 @@
         <Text is="strong">1</Text>
 
         <Box palette="dark" padding="small">
-            <Grid {...$$props} points="1" spacing="medium" style="width:18rem;">
+            <Grid.Container {...$$props} points="1" spacing="medium" style="width:18rem;">
                 <Box palette="alert" style="height:3rem;" />
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
@@ -292,7 +291,7 @@
                 <Box palette="affirmative" style="height:3rem;" />
                 <Box palette="negative" style="height:3rem;" />
                 <Box palette="alert" style="height:3rem;" />
-            </Grid>
+            </Grid.Container>
         </Box>
     </div>
 </Stack>