+++ src/components/layouts/grid/stories/grid.stories.js
@@ -6,15 +6,16 @@ import {
     STORYBOOK_SPACING_ARGUMENTS,
 } from "../../../../storybook/arguments";
 
-import Grid from "../Grid.svelte";
+import GridContainer from "../GridContainer.svelte";
 
 import GridDefaultStory from "./GridDefaultStory.svelte";
 import GridPointsStory from "./GridPointsStory.svelte";
 import GridSpacingStory from "./GridSpacingStory.svelte";
+import GridSpanStory from "./GridSpanStory.svelte";
 
 export default {
     title: "Layouts/Grid",
-    component: Grid,
+    component: GridContainer,
     argTypes: {
         ...STORYBOOK_POINTS_ARGUMENTS,
         ...STORYBOOK_SPACING_ARGUMENTS,
@@ -40,6 +41,13 @@ export const Points = (args) => ({
 
 export const Spacing = (args) => ({
     Component: GridSpacingStory,
+    props: {
+        ...args,
+    },
+});
+
+export const Span = (args) => ({
+    Component: GridSpanStory,
     props: {
         ...args,
     },